structural: extract WatchStatusService from WatchProgressService
Split read-path methods (getProgress, getWatchStatusFor*) into a
dedicated WatchStatusService with 4 dependencies, leaving the
write-path (reportTimeline, resetProgress) in WatchProgressService
with its original 7 dependencies. Both data loaders and their tests
now wire against the lighter service.
